# permission-system-client/src/main/resources/api.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/permission/user/role/remove/{roleId}/user/{uuid}": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "removeRole",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission/user/role/add/{roleId}/user/{uuid}": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "addRole",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission/update/": {
      "post": {
        "tags": [
          "permission-controller"
        ],
        "summary": "Update an existing permission.",
        "description": "This endpoint is used to update an already existing permission.",
        "operationId": "updatePermission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionInfoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Permission updated successfully.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionInfoDto"
                }
              }
            }
          },
          "404": {
            "description": "The referenced parent does not exist.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusException"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission/save/": {
      "post": {
        "tags": [
          "permission-controller"
        ],
        "summary": "Create a new permission.",
        "description": "This endpoint is used to create a new permission.",
        "operationId": "createPermission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionInfoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Permission created successfully.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionInfoDto"
                }
              }
            }
          },
          "404": {
            "description": "The referenced parent does not exist.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusException"
                }
              }
            }
          },
          "407": {
            "description": "The ID is not unique.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusException"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission/role/{id}/weight/{weight}": {
      "post": {
        "tags": [
          "role-controller"
        ],
        "operationId": "changeWeight",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "weight",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission/role/{id}/permission/{uuid}/{value}": {
      "post": {
        "tags": [
          "role-controller"
        ],
        "operationId": "addPermission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "value",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "NEUTRAL",
                "GRANTED",
                "DENIED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission/role/create": {
      "post": {
        "tags": [
          "role-controller"
        ],
        "operationId": "createRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission/user/effective/{uuid}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getEffectiveUserModelDto",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EffectiveUserModelDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission/user/effective/extensive/{uuid}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getExtensiveEffectiveUserModelDto",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensiveEffectiveUserModelDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/permission/has/{user}/permission/{permission}": {
      "get": {
        "tags": [
          "permission-controller"
        ],
        "summary": "Check if user has permission.",
        "description": "This endpoint checks if a specified user has a certain permission.",
        "operationId": "hasPermission",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "description": "The UUID of the user",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "permission",
            "in": "path",
            "description": "The UUID of the permission",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/permission/role/delete/{id}": {
      "delete": {
        "tags": [
          "role-controller"
        ],
        "operationId": "deleteRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permission/delete/{uuid}/mode/{mode}": {
      "delete": {
        "tags": [
          "permission-controller"
        ],
        "summary": "Delete permission.",
        "description": "This endpoint is used to delete a permission. There are 3 different modes: Shallow simply deletes the permission, at the risk of ripping a whole in the permission tree.Intelligent fixes that hole, by connecting all children of the deleted permission to the deleted permissions parent. Recursive also deletes all sub-permissions.",
        "operationId": "deletePermission",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "The UUID of the permission",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "mode",
            "in": "path",
            "description": "The mode used for deleting the permission.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SHALLOW",
                "INTELLIGENT",
                "RECURSIVE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission deleted successfully.",
            "content": {
              "*/*": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The referenced permission does not exist.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusException"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PermissionDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "format": "uuid"
          },
          "permissionValue": {
            "type": "string",
            "enum": [
              "NEUTRAL",
              "GRANTED",
              "DENIED"
            ]
          }
        }
      },
      "RoleDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "integer",
            "format": "int64"
          },
          "weight": {
            "type": "integer",
            "format": "int32"
          },
          "permissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionDto"
            }
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            }
          },
          "explicitPermissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionDto"
            }
          }
        }
      },
      "PermissionInfoDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "parent": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "EffectiveUserModelDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "format": "uuid"
          },
          "permissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionDto"
            }
          }
        }
      },
      "ExtensiveEffectiveUserModelDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "format": "uuid"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            }
          },
          "permissionDtos": {
            "type": "array",
            "properties": {
              "empty": {
                "type": "boolean"
              }
            },
            "items": {
              "$ref": "#/components/schemas/ExtensivePermissionDto"
            }
          }
        }
      },
      "ExtensivePermissionDto": {
        "type": "object",
        "properties": {
          "permissionDto": {
            "$ref": "#/components/schemas/PermissionDto"
          },
          "permissionOrigin": {
            "type": "string",
            "enum": [
              "SPECIFIC",
              "ROLE"
            ]
          },
          "roleDto": {
            "$ref": "#/components/schemas/RoleDto"
          }
        }
      }
    }
  }
}